+++ CHANGELOG.md
@@ -11,6 +11,11 @@
   - In the tool window, the root element's name reflects whether Gherkin and/or JBehave Story files are available in the
   project.
 
+#### ENHANCEMENT
+- The tree model in the Tags tool window is rebuilt when either the application or the project-level mapping
+has been changed in Settings. This aims to provide a smoother user experience by not having to restart the IDE or
+turn to other workarounds to reload the tool window, so that it displays data with the up-to-date mappings.
+
 ### 0.1.0
 
 Initial release